Add point struct for positions
And switch to calling bracket_lib's Point "BPoint".
It's my own struct for a similar reason to color - I don't want to be too attached to them in case I switch off.
The struct can be used for velocity too, but I didn't make that change in this commit.
I also didn't touch the map generation. Going to let that sit for now until it's time to work on level generation.