First pass at porting the game to ggez
A number of bugs and hacks, but it builds and is technically playable.
Also I didn't remove bracket's implementation of Dijkstra's - that doesn't pull in the dep I don't want and I don't want to go digging for a different implementation until I have the main port over and done with.