Move profile data to the DB and split timing into segments
The finer grained timing is useful for analysis.
The DB is partly for consistency and partly because it is not a good format for ad-hoc fields like the split names.
I'd need to switch to this same data model or limit my flexibility regarding what systems can be timed + make implementation more difficult.