Remove homing AI
The implementation was bothering me in a lot of ways, and the specific blocker was that the implementation depends on the `CONSOLE_WIDTH and HEIGHT` fields that should be about the viewport.
It was fixable, but I'm not going to until I decide that I actually want entities with full shortest path movement at any distance.
Also random is 5ms faster on my computer.