Split spawning into attempts and placement
That is, we create a new table SpawnAttempt, and objects are placed into that table and then spawned as part of a system rather than being directly spawned in.
This saves a huge amount of time (60ms to 5ms for spawning on my computer), because the previous approach sorted and then discarded the list of all spawnable squares for each entity being spawned.
I also use the opportunity to upgrade clicking to place particles on the location the player clicks instead of somewhere random.